Support setting username of username and password
Sometimes when using basic auth only the username will be set
while at other times botht the username and password is set.
The proxy should use the username as the token only when
the password component is not set. Otherwise it should
use the password to authenticate.

Update to Golang 1.16